Unprojected datasets with 2D arrays of lon and lat may now be used as…

+++ opendrift/readers/reader_netCDF_CF_generic.py
@@ -476,8 +476,8 @@ class Reader(StructuredReader, BaseReader):
         else:
             variables['x'] = indx
             variables['y'] = indy
-        variables['x'] = np.asarray(variables['x'])
-        variables['y'] = np.asarray(variables['y'])
+        variables['x'] = np.asarray(variables['x'], dtype=np.float32)
+        variables['y'] = np.asarray(variables['y'], dtype=np.float32)
 
         variables['time'] = nearestTime
